Adding security rule to protect api resources
