+++ src/main/java/pazuzu/Application.java
@@ -8,21 +8,9 @@ import org.springframework.context.annotation.ComponentScan;
 import org.springframework.stereotype.Component;
 
 @EntityScan(basePackages = "pazuzu.model")
-@ComponentScan(basePackages = {"pazuzu.service", "pazuzu.dao"})
+@ComponentScan(basePackages = {"pazuzu.service", "pazuzu.dao", "pazuzu.config"})
 @EnableAutoConfiguration
 public class Application {
-
-    @Component
-    @SuppressWarnings("unused")
-    public static class JerseyConfig extends ResourceConfig {
-        public JerseyConfig() {
-            packages("pazuzu.web");
-            registerClasses(UnknownExceptionMapper.class);
-            registerClasses(ServiceExceptionMapper.class);
-            registerClasses(AnyExceptionMapper.class);
-        }
-    }
-
     public static void main(final String[] args) {
         new SpringApplication(Application.class).run(args);
     }
