+++ src/main/resources/config/application.yml
@@ -1,5 +1,6 @@
 tokens:
   access-token-uri: https://token.services.auth.zalando.com/oauth2/access_token?realm=/services
+  token-info-uri: https://info.services.auth.zalando.com/oauth2/tokeninfo
   credentials-directory: ${user.dir}/credentials
   token-configuration-list:
    - token-id: kio
