+++ src/test/java/componenttest/AbstractComponentTest.java
@@ -1,5 +1,6 @@
 package componenttest;
 
+import org.junit.Ignore;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.SpringApplicationConfiguration;
